+++ src/Modules/Content/Content.js
@@ -1,6 +1,5 @@
 import '@/core'
 import '@/globals'
-import '@/tablesorter'
 
 import 'jquery.tinymce' // cannot go earlier!